+++ src/app/timeline/timelineWidget.js
@@ -121,7 +121,7 @@ function timelineWidgetFactory($rootScope, cfg, tokens, vis, utl) {
     addSelectListener();
 
     $rootScope.$on("tokenGeometryUpdated", function(e, token_id, geometry) {
-        console.warn("$on tokenGeometryUpdated: token_id=", token_id, "geometry=", geometry);
+        //console.warn("$on tokenGeometryUpdated: token_id=", token_id, "geometry=", geometry);
         var item = items.get(token_id);
         item.geometry = geometry;
         items.update(item);