+++ ChangeLog.md
@@ -1,6 +1,7 @@
 - 2015-04-22: (branch: geom-ol3)
-  - trigger token modification at the end of edit action: drag, add.
+  - trigger token modification at the end of edit action: drag, add.  TODO: modify.
   - tooltips
+  - add delete interaction based on select interaction plus click listener
 
 - 2015-04-21: (branch: geom-ol3)
   - improved geometry editing control via main radio-buttons: View, Move, Modify, Add.